% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_markov.R
\name{ce_markov}
\alias{ce_markov}
\title{Cost-effectiveness Markov model}
\usage{
ce_markov(
  start_pop,
  n_treat,
  p_mortality,
  state_c_matrix,
  trans_c_matrix,
  state_q_matrix,
  n_cycles = 46,
  init_age = 55,
  lambda = 0.5,
  s_names = NULL,
  t_names = NULL
)
}
\arguments{
\item{start_pop}{}

\item{n_treat}{}

\item{p_mortality}{}

\item{state_c_matrix}{}

\item{trans_c_matrix}{}

\item{state_q_matrix}{}

\item{n_cycles}{}

\item{init_age}{}

\item{lambda}{}

\item{s_names}{}

\item{t_names}{}
}
\description{
Cost-effectiveness Markov model
}
