% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_matrix_cycle.R
\name{p_matrix_cycle}
\alias{p_matrix_cycle}
\title{Probability matrix per cycle}
\usage{
p_matrix_cycle(
  age,
  cycle,
  p_mortality,
  tpProg = 0.01,
  tpDcm = 0.15,
  effect = 0.5,
  lambda = 0.5
)
}
\arguments{
\item{age}{}

\item{cycle}{}

\item{p_mortality}{Function}

\item{tpProg}{Probability}

\item{tpDcm}{Probability}

\item{effect}{Impact of drug intervention. Proportion reduction in Asymptomatic
disease to Progressive disease transition probability.}

\item{lambda}{Proportion to Progressive disease. How to redistribute individuals
who no longer transition to death when changing from Williams probabilities
to alternative. They can either go to Progressive disease or stay in
Asymptomatic disease. 0<= lambda <= 1.}
}
\description{
Probability matrix per cycle
}
