% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_mortality.R
\name{p_mortality_long}
\alias{p_mortality_long}
\title{probability of death long look-up table}
\usage{
p_mortality_long(filename = "tpDn_long.RData", baseyear = NA)
}
\description{
probability of death long look-up table
}
