% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_mortality.R
\name{p_mortality_wide}
\alias{p_mortality_wide}
\title{probability of death wide look-up table}
\usage{
p_mortality_wide(filename = "tpDn_wide.RData", baseyear = NA)
}
\description{
to use with functional data analysis
}
