% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_mortality.R
\name{p_mortality_williams}
\alias{p_mortality_williams}
\title{probability of death look-up
from Williams paper}
\usage{
p_mortality_williams(age, ...)
}
\description{
probability of death look-up
from Williams paper
}
