% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_surv.R
\name{partition_surv}
\alias{partition_surv}
\title{Fit a partitioned survival model}
\usage{
partition_surv(..., time = NULL)
}
\arguments{
\item{...}{Fitted flexsurvreg obj}

\item{time}{numeric vector of time to estimate probabilities}
}
\value{
data frame of times and probabilities for each state
}
\description{
order of fit object is from top to bottom i.e.
least nested survival curve to most e.g.
OS, PFS
}
