% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_surv.R
\name{partition_surv}
\alias{partition_surv}
\title{Fit a partitioned survival model}
\usage{
partition_surv(fit.pfs, fit.os, title = "trt", time = times)
}
\arguments{
\item{fit.pfs}{flexsurv obj fitting pfs}

\item{fit.os}{flexsurv obj fitting os}

\item{time}{numeric vector of time to estimate probabilities}
}
\value{
list w/ one entry of a data frame w/ probabilities
associated w/ stable ,prog and dead.
}
\description{
Fit a partitioned survival model
}
